RSMS-537: Refactor validation of Use Log entry form
PI use-log entries need to be constrained by total available 'activity'
as well as dates. When editing an existing (and active) Use Log entry,
the available activity needs to reflect what is already tracked in the
log (example - with a remainder of 2 mCi, editing an existing use log of
3 mCi should reflect a total available activity of 5 mCi).

Additionally, the Use Log table control for displaying active/inactive
use logs was incomplete.

When activating a use log, it needs to be validated before activation is
allowed.