RSMS-613: Fix Scint and Carboy cycle definitions to include loadDrum
The individual class files for these are deprecated and are no longer
used. Instead, the bundled version needs to be updated.

I've included updates to the individual files for consistency, and in
the hopes that the bundle can be removed from source control and instead
generated during a build process.