RSMS-957: Add debug logging in datastoreManager
The DSM does not update data which it alreay has stored. If it finds a
key exists when instructed to store something, it ignores the incoming
value. It is assumed that updates are made by requesting the existing
value and modifying it.

This commit adds logs to state that a value is being ignored, as well as
when a storage type is created and when a new value is initially stored