RSMS-1041: Prevent creation of duplicate PIs when saving User details
The saveUser endpoint looks at an incoming 'PrincipalInvestigator' field
to determine if changes to Department assignment(s) are required. If
this object was sent without a Key_id, then it would create a new PI
object even if one already existed in the database.

To resolve this, any incoming PI entry has its key_id overwritten with
that of the pre-existing one for that user. This will resolve the bug
without modifying the complexity of the endpoint.