RSMS-1013, RSMS-1090: Replace locationHub scroll-table directive with CSS sticky element
This allows us to remove the fixed elements and stop recalculating table widths constantly - we can just let the browser do it. Because of the stacked nature of
these table headers, there are still locationHub-specific height values involved in specifying the sticky-headers positions, in part due to the fact that position:sticky
will not work at a <thead> or <tr> level, and must be applied to the <th> elements. Sticky items should stack if they are all within the same container, but this cannot be the case
with our tables.