RSMS-1139: Add Report column to MyLab equipment table. Populate equipment details by mixing New and Old inspection data.
Equipment information is highly contextual, and it is imperative that we extract details from the correct point in time. Status must come from the most-recent non-future inspection, Report must come from the most-recent-with-a-report, Location information must come from most-recent, etc.

It would be wise to refactor the Equipment models so that its context can be collected without considering so many special cases. Displaying the correct information here requires a high amount of coordination between Data, Controller, and View elements.