RSMS-1229: Implement basic navigation badges for displaying numeric values.
An array of named objects ({name,count}) can be bound to the hubBannerNav directive which will match them by name to navigation items. If a matched item's 'count' is non-zero, then it will display a badge on the navigation item (with a maximum value of 9).

Since the array is bound, the provider is responsible for making any meaningful updates to these numbers